Only update the content sync timestamp after a sync that carried content.
We also use the time the sync was started as the reference point instead of the time it finished as new content might have been added/modified during the sync. This ensures that content modified half-way through the sync will be synced the next time the sync is run and not lost halfway through.