Fix white page flash.
The flash was due to the browser rendering the page and not blocking the javascript execution for the vs-region-loading attribute to be added to the header. We’re always setting the attribute in the header, this has the “minor” inconvenience thaat the page will not be “visible” for browsers with JavaScript disabled, but then again, if they have JavaScript disable they won’t be able to buy tickets or use most of the functionality of the website.